Move to packages folder and add copyright
