+++ packages/botDeploymentTask/botDeployment/package.json
