+++ packages/botDeploymentTask/images/extension-icon.png
