--- vss-extension.json
@@ -1,44 +1,0 @@
-{
-    "manifestVersion": 1,
-     "id": "bot-deployment",
-     "name": "Bot Deployment",
-     "version": "0.8.11",
-     "publisher": "Southworks-Pipelines-Test",
-     "targets": [
-         {
-             "id": "Microsoft.VisualStudio.Services"
-         }
-     ],    
-    "description": "Deploy a bot to Azure and connect it with channels",
-    "categories": [
-         "Azure Pipelines"
-     ],
-    "icons": {
-         "default": "images/extension-icon.png"        
-     },
-     "content": {
-         "details": {
-             "path": "readme.md"
-         }
-     },
-    "files": [
-         {
-             "path": "botDeployment"
-         }
-     ],
-    "contributions": [
-         {
-             "id": "custom-bot-deployment-task",
-             "type": "ms.vss-distributed-task.task",
-             "targets": [
-                 "ms.vss-distributed-task.tasks"
-             ],
-             "properties": {
-                 "name": "botDeployment"
-             }
-     }
-     ],
-     "galleryFlags": [
-         "Public"
-     ]
-}
